move model functions around
